moved generate_profile_description_missing_notifications from liberapay/billing/cron.py to liberapay/models/particpant.py, refactored dependencies #2013
