+++ tests/py/test_cron.py
@@ -8,7 +8,8 @@ from liberapay.cron import Daily, Weekly
 from liberapay.i18n.currencies import fetch_currency_exchange_rates
 from liberapay.models.participant import send_account_disabled_notifications
 from liberapay.testing.emails import EmailHarness
-from liberapay.billing.cron import generate_profile_description_missing_notifications
+from liberapay.models.participant import \
+    generate_profile_description_missing_notifications
 from liberapay.billing.payday import Payday
 from liberapay.testing import EUR
